Update 1. Repeating Element Count
O(Log N). Since it is already sorted ,so by modifying the binary search algorithm. 
Idea is - to find the index of first and last occurrence of the given key and return one more than the difference between these two index. 
for this if value is at mid and value just before mid is smaller than x,